Move UI components to separate files and add previews
