+++ chess/UI/ChessBoardView.swift
@@ -1,24 +1,11 @@
 //
-//  ContentView.swift
+//  ChessBoardView.swift
 //  Chess
 //
-//  Created by Yeliena Khaletska on 08.07.2024.
+//  Created by Andrius Shiaulis on 11.07.2024.
 //
 
 import SwiftUI
-
-struct ContentView: View {
-
-    @ObservedObject var chessModel: ChessModel
-
-    var body: some View {
-        ChessBoardView(chessModel: self.chessModel)
-            .onAppear(perform: {
-                self.chessModel.createNewGameBoard()
-            })
-    }
-
-}
 
 struct ChessBoardView: View {
 
@@ -75,24 +62,6 @@ struct ChessBoardView: View {
 
 }
 
-struct ChessPieceView: View {
-
-    private let piece: ChessPiece
-    private let imageSize: CGFloat
-
-    init(piece: ChessPiece, imageSize: CGFloat) {
-        self.piece = piece
-        self.imageSize = imageSize
-    }
-
-    var body: some View {
-        Image(piece.color.rawValue + piece.kind.rawValue)
-            .resizable()
-            .frame(width: self.imageSize, height: self.imageSize, alignment: .center)
-    }
-
-}
-
-#Preview {
-    ContentView(chessModel: ChessModel())
+#Preview(traits: .fixedLayout(width: 300, height: 300)) {
+    ChessBoardView(chessModel: .init())
 }
